REST routes

+++ services_backend/routes/button.py
@@ -2,18 +2,20 @@ from fastapi import HTTPException, APIRouter
 from fastapi_sqlalchemy import db
 
 from .models.button import ButtonCreate, ButtonUpdate, ButtonGet
+from .models.category import CategoryGet
 from ..models.database import Button, Category
 
 button = APIRouter()
 
 
 @button.post("/", response_model=ButtonGet)
-def create_button(button_inp: ButtonCreate):
-    category = db.session.query(Category).filter(Category.id == button_inp.category_id).one_or_none()
+def create_button(button_inp: ButtonCreate, category_id: int):
+    category = db.session.query(Category).filter(Category.id == category_id).one_or_none()
     if not category:
         raise HTTPException(status_code=404, detail="Category does not exist")
     last_button = db.session.query(Button).order_by(Button.order.desc()).first()
     button = Button(**button_inp.dict())
+    button.category_id = category_id
     if last_button:
         button.order = last_button.order + 1
     db.session.add(button)
@@ -21,24 +23,44 @@ def create_button(button_inp: ButtonCreate):
     return button
 
 
-@button.get("/", response_model=list[ButtonGet])
-def get_buttons(offset: int = 0, limit: int = 100):
-    return db.session.query(Button).order_by(Button.order).offset(offset).limit(limit).all()
+@button.get("/", response_model=CategoryGet, response_model_exclude_unset=True)
+def get_buttons(category_id: int):
+    category = db.session.query(Category).filter(Category.id == category_id).one_or_none()
+    if not category:
+        raise HTTPException(status_code=404, detail="Category does not exist")
+    return category
 
 
 @button.get("/{button_id}", response_model=ButtonGet)
-def get_button(button_id: int):
+def get_button(button_id: int, category_id: int):
+    category = db.session.query(Category).filter(Category.id == category_id).one_or_none()
+    if not category:
+        raise HTTPException(status_code=404, detail="Category does not exist")
     button = db.session.query(Button).filter(Button.id == button_id).one_or_none()
     if not button:
         raise HTTPException(status_code=404, detail="Button does not exist")
-    return button
+    if button.category_id != category_id:
+        raise HTTPException(status_code=404, detail="Button is not this category")
+    return {
+        "id": button_id,
+        "order": button.order,
+        "category": category_id,
+        "name": button.name,
+        "icon": button.icon,
+        "link": button.link,
+        "type": button.type,}
 
 
 @button.delete("/{button_id}", response_model=None)
-def remove_button(button_id: int):
+def remove_button(button_id: int, category_id: int):
+    category = db.session.query(Category).filter(Category.id == category_id).one_or_none()
+    if not category:
+        raise HTTPException(status_code=404, detail="Category does not exist")
     button = db.session.query(Button).filter(Button.id == button_id).one_or_none()
     if not button:
         raise HTTPException(status_code=404, detail="Button does not exist")
+    if button.category_id != category_id:
+        raise HTTPException(status_code=404, detail="Button is not this category")
     db.session.delete(button)
     db.session.query(Button) \
         .filter(Button.order > button.order) \
@@ -47,13 +69,19 @@ def remove_button(button_id: int):
 
 
 @button.patch("/{button_id}", response_model=ButtonGet)
-def update_button(button_inp: ButtonUpdate, button_id: int):
-    button = db.session.query(Button).filter(Button.id == button_id)
-    last_button = db.session.query(Button).order_by(Button.order.desc()).first()
+def update_button(button_inp: ButtonUpdate, button_id: int, category_id: int):
+    button = db.session.query(Button).filter(Button.category_id == category_id).filter(Button.id == button_id)
+    last_button = db.session.query(Button).filter(Button.category_id == category_id).order_by(Button.order.desc()).first()
+    category = db.session.query(Category).filter(Category.id == category_id).one_or_none()
+
+    if not category:
+        raise HTTPException(status_code=404, detail="Category does not exist")
     if not button.one_or_none():
         raise HTTPException(status_code=404, detail="Button does not exist")
     if not any(button_inp.dict().values()):
         raise HTTPException(status_code=400, detail="Empty schema")
+    if button.one().category_id != category_id:
+        raise HTTPException(status_code=404, detail="Button is not this category")
     if last_button and (button_inp.order > last_button.order + 1):
         raise HTTPException(status_code=400, detail=f"Can`t create button with order {button_inp.order}. "
                                                     f"Last category is {last_button.order}")
